Move TreeViewStateRestorer from Common/Controls into Common/SystemUtil
