+++ pwiz_tools/Shared/Common/Common.csproj
@@ -99,7 +99,6 @@
     <Compile Include="Controls\LiteDropDownList.cs">
       <SubType>Component</SubType>
     </Compile>
-    <Compile Include="Controls\TreeViewStateRestorer.cs" />
     <Compile Include="DataAnalysis\Clustering\ClusterResults.cs" />
     <Compile Include="DataAnalysis\Clustering\DendrogramData.cs" />
     <Compile Include="Controls\Clustering\DendrogramFormat.cs" />
