+++ pwiz_tools/Shared/Common/CommonUtil.csproj
@@ -156,6 +156,7 @@
     <Compile Include="SystemUtil\TempDir.cs" />
     <Compile Include="SystemUtil\TrackAttribute.cs" />
     <Compile Include="SystemUtil\CommonActionUtil.cs" />
+    <Compile Include="SystemUtil\TreeViewStateRestorer.cs" />
     <Compile Include="SystemUtil\ValueCache.cs" />
     <Compile Include="Collections\ValueSet.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
