+++ pwiz_tools/Shared/Common/SystemUtil/TreeViewStateRestorer.cs
@@ -16,6 +16,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 using System;
 using System.Collections;
 using System.Collections.Generic;
@@ -23,7 +24,7 @@ using System.Linq;
 using System.Text;
 using System.Windows.Forms;
 
-namespace pwiz.Common.Controls
+namespace pwiz.Common.SystemUtil
 {
     public class TreeViewStateRestorer
     {
