+++ pwiz_tools/Skyline/Controls/AuditLog/AuditLogForm.zh-CHS.resx
