+++ pwiz_tools/Skyline/Controls/TreeViewMSStateRestorer.cs
@@ -22,6 +22,7 @@ using System.Linq;
 using System.Text;
 using System.Windows.Forms;
 using pwiz.Common.Controls;
+using pwiz.Common.SystemUtil;
 
 namespace pwiz.Skyline.Controls
 {
